+++ libraries/sonars/Sonar_lib.cpp
@@ -1,5 +1,5 @@
 /**************************************************
- * Just support version > 1.0.0
+ * Just support version >= 1.0.0
  * ************************************************/
 
 #include "Sonar_lib.h"
@@ -7,21 +7,23 @@
 
 unsigned int trigPin;
 unsigned int echoPin;
+
 long duration;
-int distance;
+long distance;
 
 void Sonar::setSonarPin(unsigned int trigPin, unsigned int echoPin){
     pinMode(trigPin, OUTPUT); 
     pinMode(echoPin, INPUT); 
 }
 
-int Sonar::sonarStart(){
+long Sonar::sonarStart(){
     digitalWrite(trigPin, LOW);
     delayMicroseconds(2);
     digitalWrite(trigPin, HIGH);
     delayMicroseconds(10);
     digitalWrite(trigPin, LOW);
     duration = pulseIn(echoPin, HIGH);
+
     distance= duration*0.034/2;
     return distance;
 }
